Fix variable names and add dryrun

+++ bootstrap.sh
@@ -2,7 +2,12 @@
 
 SCRIPT_DIR=$(cd $(dirname $0) && pwd)
 SOURCE_DIR=$SCRIPT_DIR/config
-TARGET_DIR=$SCRIPT_DIR/bin
+BIN_DIR=$SCRIPT_DIR/bin
+TARGET_DIR=$HOME
+
+MANIFEST='manifest'
+
+dryrun=false
 
 function green {
   echo "\033[32m$1\033[0m"
@@ -31,7 +36,9 @@ function symlink {
     echo -e "Skipping existing symlink $(yellow $target)"
   else
     echo -e "Linking symlink $(green $target)"
-    ln -s $source $target
+    if ! $dryrun; then
+      ln -s $source $target
+    fi
   fi
 }
 
@@ -55,12 +62,12 @@ function rm_symlink {
 function install {
   echo -e $(green " --- Symlinking Configuration Files --- ")
   while read file; do
-    symlink $SOURCE_DIR/$file $HOME/.$file
+    symlink $SOURCE_DIR/$file $TARGET_DIR/.$file
   done < manifest
 
   echo -e $(green " --- Symlinking Binary Files --- ")
-  for file in $(ls $TARGET_DIR); do
-    symlink $TARGET_DIR/$file $HOME/bin/$file
+  for file in $(ls $BIN_DIR); do
+    symlink $BIN_DIR/$file $TARGET_DIR/bin/$file
   done
   echo -e "Install finished."
 }
@@ -69,12 +76,12 @@ function install {
 function uninstall {
   echo -e $(green " --- Removing Configuration Symlinks --- ")
   while read file; do
-    rm_symlink $HOME/.$file
+    rm_symlink $TARGET_DIR/.$file
   done < manifest
 
   echo -e $(green " --- Removing Binary Symlinks --- ")
-  for file in $(ls $TARGET_DIR); do
-    rm_symlink $HOME/bin/$file
+  for file in $(ls $BIN_DIR); do
+    rm_symlink $TARGET_DIR/bin/$file
   done
   echo -e "Uninstall finished. Feel cleaner?"
 }
